requirements.txt
neo4j-driver
weaviate-client
sentence-transformers
fastapi
uvicorn[standard]
groq
python-dotenv
authlib
httpx
itsdangerous
python-jose[cryptography]
